Split out common functions into tacp.py in module_utils and add ability to provide a static MAC address to VNICs
