+++ lib/ansible/modules/tacp/tacp_instance.py
@@ -3,9 +3,9 @@
 # Copyright: (c) 2020, Lenovo
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from uuid import UUID, uuid4
 from ansible.module_utils.basic import AnsibleModule
-import re
+from ansible.module_utils.tacp import *
+
 import json
 import tacp
 import sys
@@ -40,10 +40,10 @@ options:
         required: false
 
 extends_documentation_fragment:
-    - azure
+    - tacp
 
 author:
-    - Your Name (@yourhandle)
+    - Xander Madsen (@xmadsen)
 '''
 
 EXAMPLES = '''
@@ -131,172 +131,12 @@ def run_module():
     configuration.api_key['Authorization'] = module.params['api_key']
     api_client = tacp.ApiClient(configuration)
 
-    def fail_with_reason(reason):
-        result['failed'] = True
-        result['failure_reason'] = reason
-        module.exit_json(**result)
-        return
-
-    def convert_memory_abbreviation_to_bytes(value):
-        """Validate memory argument. Returns the memory value in bytes."""
-        MEMORY_RE = re.compile(
-            r"^(?P<amount>[0-9]+)(?P<unit>t|tb|g|gb|m|mb|k|kb)?$")
-
-        matches = MEMORY_RE.match(value.lower())
-        if matches is None:
-            raise ValueError(
-                '%s is not a valid value for memory amount' % value)
-        amount_str, unit = matches.groups()
-        amount = int(amount_str)
-        amount_in_bytes = amount
-        if unit is None:
-            amount_in_bytes = amount
-        elif unit in ['k', 'kb']:
-            amount_in_bytes = amount * 1024
-        elif unit in ['m', 'mb']:
-            amount_in_bytes = amount * 1024 * 1024
-        elif unit in ['g', 'gb']:
-            amount_in_bytes = amount * 1024 * 1024 * 1024
-        elif unit in ['t', 'tb']:
-            amount_in_bytes = amount * 1024 * 1024 * 1024 * 1024
-
-        return amount_in_bytes
-
-    def is_valid_uuid(uuid_to_test, version=4):
-        """
-        Check if uuid_to_test is a valid UUID.
-
-        Parameters
-        ----------
-        uuid_to_test : str
-        version : {1, 2, 3, 4}
-
-        Returns
-        -------
-        `True` if uuid_to_test is a valid UUID, otherwise `False`.
-
-        Examples
-        --------
-        >>> is_valid_uuid('c9bf9e57-1685-4c89-bafb-ff5af830be8a')
-        True
-        >>> is_valid_uuid('c9bf9e58')
-        False
-        """
-        try:
-            uuid_obj = UUID(uuid_to_test, version=version)
-        except ValueError:
-            return False
-
-        return str(uuid_obj) == uuid_to_test
-
-    def get_component_fields_by_name(name, component, api_client, fields=['name', 'uuid']):
-        """
-        Returns the UUID of a named component if it exists in a given 
-        ThinkAgile CP cloud, otherwise return None.
-
-        :param name The name of the component that may or may not exist yet
-        :type name str
-        :param component The type of component in question, must be one of 
-        """
-
-        valid_components = ["storage_pool", "application",
-                            "template", "datacenter", "migration_zone",
-                            "vnet", "vlan"]
-
-        if component not in valid_components:
-            return "Invalid component"
-        if component == "storage_pool":
-            api_instance = tacp.FlashPoolsApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_flash_pools_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_flash_pools_using_get: %s\n" % e
-        elif component == "application":
-            api_instance = tacp.ApplicationsApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_applications_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_applications_using_get: %s\n" % e
-        elif component == "template":
-            api_instance = tacp.TemplatesApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_templates_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_templates_using_get: %s\n" % e
-        elif component == "datacenter":
-            api_instance = tacp.DatacentersApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_datacenters_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_datacenters_using_get: %s\n" % e
-        elif component == "migration_zone":
-            api_instance = tacp.MigrationZonesApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_migration_zones_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_migration_zones_using_get: %s\n" % e
-        elif component == "vlan":
-            api_instance = tacp.VlansApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_vlans_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_vlans_using_get: %s\n" % e
-        elif component == "vnet":
-            api_instance = tacp.VnetsApi(api_client)
-            try:
-                # View flash pools for an organization
-                api_response = api_instance.get_vnets_using_get(
-                    fields=fields)
-            except ApiException as e:
-                return "Exception when calling get_vnets_using_get: %s\n" % e
-
-        if (api_response):
-            if fields == ['name', 'uuid']:
-                for result in api_response:
-                    if result.name == name:
-                        return result.uuid                
-            if 'bootOrder' in fields:
-                for result in api_response:
-                    if result.name == name:
-                        boot_order = []
-                        for order_item in result.boot_order:
-                            str_dict = str(order_item).replace(
-                                "\n", "").replace("'", '"').replace("None", '""')
-
-                            json_dict = json.loads(str_dict)
-
-                            disk_uuid = json_dict['disk_uuid'] if json_dict['disk_uuid'] else None
-                            name = json_dict['name'] if json_dict['name'] else None
-                            order = json_dict['order'] if json_dict['order'] else None
-                            vnic_uuid = json_dict['vnic_uuid'] if json_dict['vnic_uuid'] else None
-                            boot_order_payload = tacp.ApiBootOrderPayload(disk_uuid=disk_uuid,
-                                                                        name=name,
-                                                                        order=order,
-                                                                        vnic_uuid=vnic_uuid)
-
-                            #sys.stdout.write(str(boot_order_payload) + '\n')
-                            boot_order.append(boot_order_payload)
-                        return boot_order
-        return None
-
     instance_params = {}
 
     # Check if VM instance with this name exists already
     instance_uuid = get_component_fields_by_name(
         module.params['name'], 'application', api_client)
-    result['instance_uuid'] = instance_uuid
+
     if instance_uuid:
         # VM does exist - so for now we need to just break without changing anything,
         # but in the future this would mean we go to edit the existing VM with this name/uuid
@@ -305,8 +145,8 @@ def run_module():
         result['changed'] = False
 
         module.exit_json(**result)
-    
-        # VM does not exist yet, so we must create it
+
+    # VM does not exist yet, so we must create it
     instance_params['instance_name'] = module.params['name']
 
     # Check if storage pool exists, it must in order to continue
@@ -357,6 +197,13 @@ def run_module():
             network_uuid = get_component_fields_by_name(
                 nic['network'], 'vlan', api_client)
 
+        if nic['mac_address']:
+            automatic_mac_address = False
+            mac_address = nic['mac_address']
+        else:
+            automatic_mac_address = True
+            mac_address = None
+
         if i == 0:
             vnic_payloads = []
             for boot_order_item in boot_order:
@@ -364,23 +211,27 @@ def run_module():
                     vnic_uuid = boot_order_item.vnic_uuid
 
         else:
-            vnic_uuid = str(uuid4())
             # TODO - his will eventually need to be modified so that any
             # boot order is possible for any extra NICs
+
+            vnic_uuid = str(uuid4())
             vnic_boot_order = len(boot_order) + i
+            
             vnic_payload = tacp.ApiAddVnicPayload(
-                automatic_mac_address=True,
+                automatic_mac_address=automatic_mac_address,
                 name=nic['name'],
                 network_uuid=network_uuid,
-                boot_order=vnic_boot_order
+                boot_order=vnic_boot_order,
+                mac_address=mac_address
             )
             vnic_payloads.append(vnic_payload)
 
         network_payload = tacp.ApiCreateOrEditApplicationNetworkOptionsPayload(
             name=nic['name'],
-            automatic_mac_assignment=True,
+            automatic_mac_assignment=automatic_mac_address,
             network_uuid=network_uuid,
             vnic_uuid=vnic_uuid,
+            mac_address=mac_address
         )
         network_payloads.append(network_payload)
 
@@ -421,11 +272,11 @@ def run_module():
             module.exit_json(**result)
 
         except ApiException as e:
-            
+
             return "Exception when calling ApplicationsApi->create_application_from_template_using_post: %s\n" % e
 
     create_instance_response = create_instance(instance_params)
-    
+
     result['create_instance_response'] = create_instance_response
 
     # use whatever logic you need to determine whether or not this module
